backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langchain-anthropic==0.1.0
langgraph==0.0.26
langchain-core==0.1.0

# Vector Database
pinecone-client==3.0.0
qdrant-client==1.7.0

# Database
supabase==2.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Image Generation
diffusers==0.25.0
transformers==4.36.0
torch==2.1.0
pillow==10.1.0
accelerate==0.25.0

# Image Analysis
opencv-python==4.8.1.78
pytesseract==0.3.10

# API Clients
tweepy==4.14.0
python-instagram==1.3.2
python-linkedin-v2==0.9.0
praw==7.7.1
google-api-python-client==2.108.0

# Trending Topics
pytrends==4.9.2
feedparser==6.0.10

# Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Monitoring
prometheus-client==0.19.0

# Production
gunicorn==21.2.0
uvloop==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
